legacy/pyproject.toml
[project]
name = "human-archive"
version = "0.1.0"
description = "AI persona archiving service - archive experts and chat with their AI personas"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "yt-dlp>=2024.0",
    "chromadb>=1.0",
    "httpx>=0.27",
    "mcp[cli]>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
human-archive-mcp = "human_archive.mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/human_archive"]
